Sample number processor variable

+++ src/LoggerDreamHost.h
@@ -47,7 +47,7 @@ public:
     // DreamHost URL and then streams out a get request
     // over that connection.
     // The return is the http status code of the response.
-    int postDataDreamHost(void);
+    int16_t postDataDreamHost(void);
 
     // This prevents the logging function from dual-posting to EnviroDIY
     void disableDualPost(void);
