+++ src/sensors/MeaSpecMS5803.h
@@ -72,7 +72,7 @@ class MeaSpecMS5803 : public Sensor
 {
 public:
     MeaSpecMS5803(int8_t powerPin, uint8_t i2cAddressHex = 0x76,
-                  int maxPressure = 14, uint8_t measurementsToAverage = 1);
+                  int16_t maxPressure = 14, uint8_t measurementsToAverage = 1);
     ~MeaSpecMS5803();
 
     bool setup(void) override;
@@ -82,7 +82,7 @@ public:
 protected:
     MS5803 MS5803_internal;
     uint8_t _i2cAddressHex;
-    int _maxPressure;
+    int16_t _maxPressure;
 };
